# backend/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Luma Local Media API
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/v1/media:
    get:
      summary: 查询媒体列表
      description: |
        默认排除状态为 missing 的媒体，以及已禁用或软删除来源下的媒体。
        sort=duration 时仅包含 duration 已稳定的行（duration_ms 非空，或 status 为 ready/failed）。
        sort=file_size 时仅包含 status 为 ready/failed 的行。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: q
          description: 按文件名搜索，最多 200 个字符。
          schema: { type: string, maxLength: 200 }
        - in: query
          name: type
          schema: { type: string, enum: [video, image] }
        - in: query
          name: library_kind
          description: 严格限制视频来源用途；个人视频页使用 personal，避免混入电影和电视剧源。图片库仅按 type=image 查询。
          schema: { type: string, enum: [personal, movies, tv] }
        - in: query
          name: favorite
          description: true 仅返回收藏媒体；false 包含未创建用户数据的媒体。
          schema: { type: boolean }
        - in: query
          name: tag_id
          description: 仅返回关联当前用户指定标签的媒体。
          schema: { type: string, maxLength: 200 }
        - in: query
          name: watch_status
          description: 按当前用户观看状态筛选；无用户数据的媒体归入 unwatched。
          schema: { type: string, enum: [unwatched, watching, completed] }
        - in: query
          name: sort
          description: created_at/filename 含处理中媒体；duration/file_size 仅含排序键已稳定的行。
          schema:
            type: string
            enum: [created_at, filename, duration, file_size]
            default: created_at
        - in: query
          name: order
          schema: { type: string, enum: [asc, desc], default: desc }
        - in: query
          name: cursor
          description: 上一页返回的不透明游标；筛选或排序条件改变后不可复用。
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: 媒体列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MediaPage" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/media/{id}/thumbnail:
    parameters:
      - $ref: "#/components/parameters/MediaID"
    get:
      summary: 获取媒体缩略图
      description: |
        返回当前可用的默认或卡片图片资产；支持使用 If-None-Match 进行 ETag 条件请求。
        若资产已记录内容哈希且 If-None-Match 命中，服务端可直接返回 304 而不读取文件体。
        内容变更重新生成完成前，仍可能提供上一份 ready 缩略图。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: variant
          description: 省略时保持原有默认缩略图行为；card 为 640×400 居中裁剪卡片图。
          schema: { type: string, enum: [default, card], default: default }
        - in: header
          name: If-None-Match
          description: 客户端缓存的缩略图 ETag。
          schema: { type: string }
      responses:
        "200":
          description: 缩略图二进制内容
          headers:
            ETag:
              description: 当前缩略图资产的强 ETag。
              schema: { type: string }
            Cache-Control:
              description: 固定为 private, max-age=604800, must-revalidate。
              schema: { type: string, const: "private, max-age=604800, must-revalidate" }
          content:
            image/*:
              schema: { type: string, format: binary }
        "304":
          description: If-None-Match 与当前 ETag 匹配，无响应体。
          headers:
            ETag:
              schema: { type: string }
            Cache-Control:
              schema: { type: string, const: "private, max-age=604800, must-revalidate" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaOrThumbnailNotFound" }
  /api/v1/media/{id}/stream:
    parameters:
      - $ref: "#/components/parameters/MediaID"
    get:
      summary: 播放原始视频
      description: |
        仅服务可见的视频原文件，支持标准单段 HTTP Range、条件请求和进度跳转。
        discovered、probing、thumbnailing、ready、failed 状态均可 Direct Play；图片和不可见媒体返回 404。
        弱 ETag 与 Content-Length 均基于打开瞬间的 size/mtime 快照；媒体源根目录不可访问时返回 503 SOURCE_OFFLINE。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: Range
          schema: { type: string, examples: ["bytes=0-1048575"] }
        - in: header
          name: If-None-Match
          schema: { type: string }
        - in: header
          name: If-Modified-Since
          schema: { type: string }
        - in: header
          name: If-Range
          description: 弱 ETag 不能用于 If-Range 强比较；客户端可使用 Last-Modified 日期。
          schema: { type: string }
      responses:
        "200":
          description: 完整视频内容
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            ETag: { schema: { type: string }, description: 由打开瞬间文件大小和秒级修改时间生成的弱 ETag。 }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
          content:
            video/*:
              schema: { type: string, format: binary }
        "206":
          description: Range 指定的部分视频内容
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            Content-Range: { schema: { type: string } }
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
          content:
            video/*:
              schema: { type: string, format: binary }
        "304":
          description: 条件请求命中，无响应体
          headers:
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaNotFound" }
        "416":
          description: Range 不可满足；Content-Range 为 bytes */快照文件大小，响应体由 Go 标准库生成。
          headers:
            Content-Range: { schema: { type: string } }
        "503":
          description: 媒体源根目录当前不可访问，error.code 为 SOURCE_OFFLINE
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    head:
      summary: 查询原始视频响应元数据
      description: 与 GET 使用相同的认证、可见性和条件请求语义，但不返回响应体。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: If-None-Match
          schema: { type: string }
        - in: header
          name: If-Modified-Since
          schema: { type: string }
      responses:
        "200":
          description: 视频响应元数据
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
        "304": { description: 条件请求命中，无响应体 }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaNotFound" }
        "503":
          description: 媒体源根目录当前不可访问，error.code 为 SOURCE_OFFLINE
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/media/{id}/original:
    parameters:
      - $ref: "#/components/parameters/MediaID"
    get:
      summary: 浏览原始图片
      description: |
        仅服务可见的图片原文件，支持 HTTP Range 和条件请求；视频和不可见媒体返回 404。
        discovered、probing、thumbnailing、ready、failed 状态均可读取原图，不依赖缩略图处理结果。
        弱 ETag 与 Content-Length 均基于打开瞬间的 size/mtime 快照；媒体源根目录不可访问时返回 503 SOURCE_OFFLINE。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: Range
          schema: { type: string, examples: ["bytes=0-1048575"] }
        - in: header
          name: If-None-Match
          schema: { type: string }
        - in: header
          name: If-Modified-Since
          schema: { type: string }
        - in: header
          name: If-Range
          description: 弱 ETag 不能用于 If-Range 强比较；客户端可使用 Last-Modified 日期。
          schema: { type: string }
      responses:
        "200":
          description: 完整原始图片
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            ETag: { schema: { type: string }, description: 由打开瞬间文件大小和秒级修改时间生成的弱 ETag。 }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
            X-Content-Type-Options: { schema: { type: string, const: nosniff } }
          content:
            image/*:
              schema: { type: string, format: binary }
        "206":
          description: Range 指定的部分图片内容
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            Content-Range: { schema: { type: string } }
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
            X-Content-Type-Options: { schema: { type: string, const: nosniff } }
          content:
            image/*:
              schema: { type: string, format: binary }
        "304":
          description: 条件请求命中，无响应体
          headers:
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
            X-Content-Type-Options: { schema: { type: string, const: nosniff } }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaNotFound" }
        "416":
          description: Range 不可满足；Content-Range 为 bytes */快照文件大小，响应体由 Go 标准库生成。
          headers:
            Content-Range: { schema: { type: string } }
        "503":
          description: 媒体源根目录当前不可访问，error.code 为 SOURCE_OFFLINE
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    head:
      summary: 查询原始图片响应元数据
      description: 与 GET 使用相同的认证、可见性和条件请求语义，但不返回响应体。
      security: [{ bearerAuth: [] }]
      parameters:
        - in: header
          name: If-None-Match
          schema: { type: string }
        - in: header
          name: If-Modified-Since
          schema: { type: string }
      responses:
        "200":
          description: 原始图片响应元数据
          headers:
            Accept-Ranges: { schema: { type: string, const: bytes } }
            Content-Length: { schema: { type: integer, format: int64 } }
            ETag: { schema: { type: string } }
            Last-Modified: { schema: { type: string } }
            Cache-Control: { schema: { type: string, const: "private, max-age=0, must-revalidate" } }
            X-Content-Type-Options: { schema: { type: string, const: nosniff } }
        "304": { description: 条件请求命中，无响应体 }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/MediaNotFound" }
        "503":
          description: 媒体源根目录当前不可访问，error.code 为 SOURCE_OFFLINE
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
...
    MediaPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/MediaSummary" }
        next_cursor:
          type: [string, "null"]
          description: 下一页的不透明游标；没有下一页时为 null。
...
    MediaSummary:
      type: object
      required: [id, title, filename, media_type, library_kind, duration_ms, width, height, thumbnail_url, card_thumbnail_url, stream_url, original_url, favorite, progress_ms, completed, last_played_at, user_data_revision, status, created_at]
      properties:
        id: { type: string }
        title: { type: string, description: 按 custom_title、detected_title、filename 顺序计算的展示标题。 }
        filename: { type: string }
        media_type: { type: string, enum: [video, image] }
        library_kind: { type: string, enum: [personal, movies, tv] }
        duration_ms: { type: [integer, "null"], format: int64, minimum: 0 }
        width: { type: [integer, "null"], minimum: 1 }
        height: { type: [integer, "null"], minimum: 1 }
        thumbnail_url:
          type: string
          description: 当前媒体的缩略图端点；不存在 ready 默认缩略图资产时为空字符串。
        card_thumbnail_url:
          type: string
          description: 16:10 居中裁剪的卡片缩略图端点；尚未补齐时为空字符串。
        stream_url:
          type: [string, "null"]
          description: 视频的 Direct Play 地址；图片固定为 null。
        original_url:
          type: [string, "null"]
          description: 图片原文件地址；视频固定为 null。
        favorite: { type: boolean, description: 读取现有用户数据；无记录时为 false。 }
        progress_ms: { type: integer, format: int64, minimum: 0, description: 读取现有用户数据；无记录时为 0。 }
        completed: { type: boolean, description: 是否已达到服务端完成阈值。 }
        last_played_at: { type: [string, "null"], format: date-time }
        user_data_revision: { type: integer, format: int64, minimum: 0 }
        status: { type: string, enum: [discovered, probing, thumbnailing, ready, failed] }
        created_at: { type: string, format: date-time, description: 媒体首次进入媒体库的时间。 }
...
  parameters:
...
    MediaID:
      in: path
      name: id
      required: true
      schema: { type: string, minLength: 1 }
  responses:
    Unauthorized:
      description: API Token 缺失或无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InvalidRequest:
      description: 请求参数、请求体或业务输入无效
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    MediaNotFound:
      description: 媒体不存在或当前不可见，error.code 为 MEDIA_NOT_FOUND
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: MEDIA_NOT_FOUND, message: 媒体不存在, details: null }
    MediaOrThumbnailNotFound:
      description: 媒体不存在时 error.code 为 MEDIA_NOT_FOUND；媒体没有可用缩略图资产时为 THUMBNAIL_NOT_FOUND
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            media_not_found:
              value:
                error: { code: MEDIA_NOT_FOUND, message: 媒体不存在, details: null }
            thumbnail_not_found:
              value:
                error: { code: THUMBNAIL_NOT_FOUND, message: 缩略图不存在, details: null }
...